+++ include/dpi.h
@@ -69,9 +69,9 @@ extern "C" {
 
 // define ODPI-C version information
 #define DPI_MAJOR_VERSION   4
-#define DPI_MINOR_VERSION   5
-#define DPI_PATCH_LEVEL     1
-#define DPI_VERSION_SUFFIX
+#define DPI_MINOR_VERSION   6
+#define DPI_PATCH_LEVEL     0
+#define DPI_VERSION_SUFFIX  "b1"
 
 #define DPI_STR_HELPER(x)       #x
 #define DPI_STR(x)              DPI_STR_HELPER(x)
@@ -681,6 +681,8 @@ struct dpiObjectTypeInfo {
     int isCollection;
     dpiDataTypeInfo elementTypeInfo;
     uint16_t numAttributes;
+    const char *packageName;
+    uint32_t packageNameLength;
 };
 
 // structure used for creating pools